improvement: clean up zod  auth
