--- libs/entities/src/schemas/auth-schemas/index.ts
@@ -1,9 +1,0 @@
-export * from "./forgot-password";
-export * from "./login";
-export * from "./logout";
-export * from "./register";
-export * from "./req-user";
-export * from "./resend-otp";
-export * from "./reset-password";
-export * from "./verify";
-export * from "./header";
