--- libs/entities/src/schemas/auth-schemas/resend-otp/index.ts
@@ -1,14 +1,0 @@
-import { z } from "zod";
-
-export const VSResendOtp = z.object({
-  email: z
-    .string()
-    .email({
-      message: "Email tidak valid",
-    })
-    .nonempty({
-      message: "Email tidak boleh kosong",
-    }),
-});
-
-export type TVSResendOtp = z.infer<typeof VSResendOtp>;
