+++ libs/entities/src/schemas/auth/login.schema.ts
@@ -15,3 +15,14 @@ export const VSLogin = z.object({
 });
 
 export type TVSLogin = z.infer<typeof VSLogin>;
+
+export const VSHeaderLogin = z.object({
+  "app-origin": z
+    .string()
+
+    .nonempty({
+      message: "Origin tidak boleh Kosong",
+    }),
+});
+
+export type TVSHeaderLogin = z.infer<typeof VSHeaderLogin>;
