+++ libs/entities/src/schemas/auth/register.schema.ts
