+++ libs/entities/src/schemas/index.ts
@@ -1,4 +1,4 @@
-export * from "./auth-schemas";
+export * from "./auth";
 export * from "./student-schemas";
 export * from "./user-schemas";
 export * from "./payment-schemas";
